fix: resolve NextAuth build error by refactoring authOptions
- Move authOptions from route file to separate src/lib/auth/auth-options.ts
- Update NextAuth route to only export required HTTP handlers (GET, POST)
- Fix import paths in all auth routes and tests to use new authOptions location
- Ensure Next.js 13+ App Router compliance by removing invalid exports from route files
- Build now compiles successfully with no TypeScript or linting errors
