+++ src/app/api/auth/force-logout/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server';
 import { getServerSession } from 'next-auth';
-import { authOptions } from '../[...nextauth]/route';
+import { authOptions } from '@/lib/auth/auth-options';
 
 // Force dynamic rendering for this route
 export const dynamic = 'force-dynamic';
